Adrian Pop, adrpo@ida.liu.se, 2006-02-14
Fixed the endcolumn and endline for all stuff we are interested in!
Check out the new interactive_api.txt.

Fixed so "Error" is not returned when classes with no parts are 
queried with getElementsInfo. Now it returns "{ }" for those.
Example: getElementsInfo(Type) => { }
Fixed getElementsInfo so it does not return class file/position 
info when the element is also a class definition.

Fixed so getCrefInfo(cref) now returns also the end line and end column
like this: filename.mo,(writable|readonly),startline,startcolumn,endline,endcolumn



git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@2111 f25d12d1-65f4-0310-ae8a-bbce733d8d8e