Moved all definitions for "Clock"-Type from Absyn.mo to DAE.mo: Probably cleaner solution since the "Clock" type is not part of the grammar.
Changed some fields in the "ClockKind" to hold values of type "DAE.Exp" as requested by Lennart.

Removed the evaluation of "Clock"-Constructors to "Clock"-Values during elaboration and instead flagged constructors as "impure". This was necessary since by moving everything to DAE.mo I suddenly had cyclic dependencies from Values.mo to DAE.mo. Although it was possible with the bootstrapped compiler to avoid a cyclic dependency error by importing DAE.mo as "protected", that didn't work for the RML compiler. This is, as I learned, because the RML compiler can not handle mutual dependencies.

git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@22311 f25d12d1-65f4-0310-ae8a-bbce733d8d8e