Add API diffModelicaFileListings
The function takes two strings and a formatter. The output is either
a diff or the resulting text after merging the Modelica strings.
The API is designed to consider the list API and indents new elements
properly, while ignoring whitespace added by list() as well as moves
comments into the correct places. Partially fixes #2905.