Move threadData away from simulation data
Each thread needs its own threadData, and must be initalized. By moving
the data, it will be easier to see if the correct threadData is passed.
This resolves issues with HPCOM using the same thread-local data for all
threads.

Move the generated function prototypes to a common header since having
the function prototypes to a common header gives compiler errors if
there is a mismatch, instead of run-time errors.