Allow generation of source-code FMUs
All generated (modelica-exchange) FMUs now include the source code of
the runtime, so that it is possible to compile an (almost) static
FMU.

By default, the runtime is still linked to the generated FMU, but
this can be changed by using -d=buildStaticSourceFMU.

The makefiles for the FMUs have been changed, such that we first
generate a source-code FMU, copying in all the files. Then a second
Makefile is used to compile the binary. Both make commands are
executed by default for translateModelFMU, but this API function will
only generate pure source-code FMUs in the future (buildModelFMU will
also compile the FMU, and there should be an API to compile
additional targets for an existing FMU).

FMUs no longer use garbage collection by default since some
simulators used libgc, and it just caused all sorts of headaches.