Use autoconf in the source-code FMUs
By using autoconf, we move a lot of the work from the user of the
Makefiles to autoconf itself.

To try it out, cross-compiling for multiple platforms (with the FMU
unzipped to M.fmutmp):

(for arch in x86_64-w64-mingw32 x86_64-linux-gnu arm-linux-gnueabihf \
    i686-linux-gnu i686-w64-mingw32
do
  (cd M.fmutmp/sources/ && \
   ./configure --host=$arch CFLAGS=-Os && \
   make -j6) || exit 1;
done)

So it is enough to pass the host and optionally CFLAGS (on Ubuntu).
Recommended packages on an x86_64 host:
apt-get install gcc-arm-linux-gnueabihf libc6-dev:i386 gcc-mingw-w64

Note that installing gcc-multilib conflicts with gcc-arm, which means
you need the i386 libc6 headers.