Store all optimization modules in only one list
Optimization modules are no longer allowed to fail. If it is intended
that a module should fail for some reason, then that need to be handled
within the module from now on.
New modules are now defined using a function pointer to the module
implementation, a name as string, and a Boolean that defines if that
module is "post-optimization only" or not.