Try to detect LAPACK/BLAS better
This includes detecting ATLAS, as well as looking for static versions
of LAPACK/BLAS that can be linked into a shared object (-fPIC). The
only such library found on Ubuntu systems seems to be OpenBLAS (but
only with -lm and -lpthread dynamically linked, since -lm -lpthread
are not compiled with -fPIC).