Susan: Do not assign the same variable twice
Earlier, Susan would assign to the same variable twice when calling
a template that takes two Text arguments and was passed the same
variable name for both of them. Now, the last argument is used in the
assignment.

Clean templates at stage 3 bootstrapping in order to get the latest
version of Susan-generated code.