Constant variable names as constant patterns
This fixes #3005 by allowing constant identifiers that evaluate to
constant literal expressions of suitable type (boxed, integer, real,
string, boolean, enumeration) to be used in a pattern. For boxed
variables, we replace the pattern expression with a shared literal
and perform valueEq on this. This should prove to be fast especially
if referenceEq holds (such as when passing a constant to a function).