Fix scalability for Tearing.selectCausalVars
Tearing.selectCausalVars does intersection of each row, which
typically has a small size (say 2), with selEqs which has size ~500
and is fixed for all rows. The indexes are in a known range (for
example 1~1000), so an array<Boolean> of the known maximum size is
created and filled with true/false (true=exists in selEqs).

This makes the intersection trivial to calculate.