Fix bottleneck for create model info
This is related to ticket:3813. Uses an array of SimVar instead of
pushing everything into a large uniontype record for every change
to a simvar. The array is indexed by an enumeration, which maintains
some knowledge of which field a particular index belongs to. The
enumeration has elements ordered in FMI order since it is important
these indexes appear in a certain order.