Support for writing Susan text directly to file
This partially resolves ticket:3356 (separators with alignment and
wrapping was not yet implemented for direct file output). It is now
possible to make a Tpl.Text bound to a file. Such a file needs to be
explicitly closed because the Susan engine does not know what it is.
It is simply a Tpl.mo function that changes the internal structure
of the Text and writes its output directly to file. This is only
possible because it knows that it cannot be read. You could write in
the template that you want to append a FILE_TEXT to a MEM_TEXT, but
this would just result in errors.

There are a few ways to use this new functionality.

From MetaModelica:

```mo
Tpl.closeFile(somTplFunc,Tpl.redirectToFile(Tpl.emptyTxt, file));
```

From Susan:

```
let &literalsFile = buffer ""
let &literalsFile += redirectToFile('<%fileNamePrefix%>_literals.h')
let &literalsFile += simulationLiteralsFile(fileNamePrefix, literals)
let &literalsFile += closeFile()
```

Susan alternative:
```
redirectToFile('<%fileNamePrefix%>_literals.h') +
simulationLiteralsFile(fileNamePrefix, literals)
closeFile()
```

Note that fully taking advantage of FILE_TEXT requires using as few let-
expressions as possible and as few impure external functions as
possible. The reason is that let-expressions need to be stored in memory
and that writing the expressions directly in the template in order is
very hard with impure functions.