Replace CREF with ComponentPrefix in ElementSource
It takes time to calculate a cref in elementsource and the only real
usage of cref is in StateMachineFeatures (but it could just as well
use hashing on the prefix; this is not done yet).

Also removed generation of the _info.xml (replaced by the JSON file
which has been the default for a long time).