Pass more linker flags to sundials cmake
Previous, CFLAGS had an override, ignoring the user CFLAGS and setting
them to -lm, which is a linker flag, giving warnings. This has been
changed to also pass the library directory which is needed for OSX.
It was also tested to properly use linker flags instead of cmake cflags,
but this break the test suite.