Improved performance of SimCode zero-crossings
Changed the functions to not consume memory when not collecting
anything. When the are many independent equation systems, the
zero-crossings are traversed once per system, making the translation
slow (this is still true; it just uses less memory now).

In order to fix the problems, it needs to be possible to query which
equation system a zero-crossing belongs to in constant time (or vice
versa).