Encapsulate the ZeroCrossings set type
Note that the relations will not use the set type since the current
implementation depends on adding duplicate relations in order to
generate correct zero-crossings. This should be changed in the future.

This commit will not improve performance, but rather abstracts away
some methods to make it simpler to experiment with hashtables and AVL
trees for representing zero-crossings.