Structural parameters as HashSet instead of Table
The FCore.Cache stores the structural parameters as a HashTable, but
does not use the value. This changes the structure to a HashSet
instead.

Note: This also fixes a bug in a model where the HashTable gets
corrupted (probably due to matchcontinue use). The structure should
possibly be changed to an AvlSet instead (or HashTable/HashSet
changed to be fully mutable).