Make test utils able to clean more generated files
This is necessary since Cppruntime statically links its runtime system
with the simulation. It caused the library testing server to have 50GB
of unused files lying around. It is less crucial for the C runtime since
the executable is always cleaned (this is an so-file for cppruntim
which was not cleaned up), and because the objects were small (few kB
instead of 10MB).