Remove equations like "{} = f(...)"
This should not be done, but breaks the Chemical library if we don't
and this was the previous behaviour of the compiler. Should be fixed
properly by patching index reduction, sorting/matching, or adding
assertions for lhs/rhs that the expressions evaluate to zero-dimension
(or at least that f(...) does not cause run-time errors).