Fix crash is default linear solver
If an expression makes a longjmp, the linear solver data was not
restored. This fix makes it so the solver data does not need to be
patched depending on which linear solver is running; it simply allocates
memory for an extra void* to hold the fallback solver's data.

Belonging to [master]:
  - OpenModelica/OMCompiler#1894