Unify reading enumeration flags
This commit simplifies the handling of the command-line arguments. It
fixed some typos from copy-paste, etc. Hopefully generated FMUs will be
able to set flags in the future as well (then the routines handling the
flags need to be moved to a separate module since they are all part of
simulation_runtime.cpp right now).

Belonging to [master]:
  - OpenModelica/OMCompiler#1908