Let the CSE look into if-conditions
If-conditions always execute, so we can consider them for CSE
optimizations. This improves some Media models by a factor of 10x since
the function calls are called inside an algebraic loop (multiple times,
although this matters little relative to moving the call outside of the
loop).

This fixes an issue with performance discovered in ticket:4423.

Belonging to [master]:
  - OpenModelica/OMCompiler#2094
  - OpenModelica/OpenModelica-testsuite#812