[FMI2] Handle the jump buffers better
The catch in try/catch needs execute, so do not return before the
jump buffers have been restored.

Also tries to detect a parent OM calling the FMU. The threadData would
be shared between the two, but the functions now set and restore the
threadData at each call if a parent (using pthread_setspecific) has
been detected.

Belonging to [master]:
  - OpenModelica/OMCompiler#2116
  - OpenModelica/OpenModelica-testsuite#829