Change the default simulation opt. level to -Os
Testing has shown that the performance of clang seems to be good enough
to enable -Os by default now. It's still slightly slower for most
models, especially small ones. But it should be a good default for
larger models.

The optimization level for simulations is now set by Susan, and
setCFlags will override this. This makes it possible for dynload to use
a different optimization level unless set by setCFlags. The disadvantage
is that getCFlags does not show what optimization level the simulations
will use (unless explicitly set by setCFlags), and this will be different
for C++ and C.

Note that this is how the MinGW version of OM worked before (except the
default for simulations was -O0).

OMEdit also has an option to change optimization level, so it should be
reasonably convenient for the user to change this to -O0 if desired.

Disabled -march=native as a default flag since it is annoying for FMUs.

Belonging to [master]:
  - OpenModelica/OMCompiler#1869