Introduce aliasing of SimCode equations
Very many equations are the same in the continuous, initial, lambda0,
and parameter equation systems. This makes some of these equations
become aliases to other ones. The aliasing is performed as the last
step before code generation since this is the most conservative
approach (it might be bad for performance since the same traversals
have been performed on the different systems). Strongly connected
components, if-equations, and algorithms are not considered yet.
Roughly 1/3 of the equations seem to be removed by this aliasing which
hopefully increases performance.

Belonging to [master]:
  - OpenModelica/OMCompiler#2361
  - OpenModelica/OpenModelica-testsuite#922