Add a Jenkins pipeline
Includes caching of omlibrary / runtest.db by using docker volumes.
This caching is done on a per-node and -branch level, so the first time
the job executes the result might be a slower-running test.

Builds using bionic/clang/qt5 and xenial/gcc/qt4, first by building and
stashing OMC and then by building the GUI and running tests in parallel.

The bibtex, MetaModelica, ParMod, Matlab, and FMU cross-compilation
tests run as separate tasks (some with additional dependencies). The FMU
cross-compilation is tested by using wine or by running the actual
ARM / Apple hardware.

Tests are ordered according to expected run-time. Nodes in the Jenkins
master have been given priority in such a way that the strongest nodes
run the most expensive testsuite runs.

Note: In the future we could distribute tests over more machines (which
helps when running on 6-core machines, but not as much on 20-core
machines).

More jobs should be moved into this one in the future (like generation
of documentation).