Disable crane C++ FMU test
The test fails randomly, even when running the exact same docker image
on the same machine, which suggests the error in not totally reliant on
the CPU architecture used, but perhaps also system load.

```
At t = 0 and h = 1e-24, the corrector convergence test failed repeatedly
or with |h| = hmin.
```

Belonging to [master]:
  - OpenModelica/OpenModelica-testsuite#1008