[Cpp] Don't use outputPath for results file, ticket:4773
The results file has its own call argument -F (or -r for the C runtime)
that is typically used with an absolute file name, including path.
A results file with relative name should be placed in the current
directory according to discussions at dev meeting on 12/03, 2018.

Moreover raise an error if open of the results file fails.

Belonging to [master]:
  - OpenModelica/OMCompiler#2813