Do not place autogenerated commands to OMC window command line history
When using the OMC command line window for debugging, it is easy to
accidentally fetch some new model, so lots of commands are issued.

Previously, all of them gone to command line history, so one not only
don't see the results of manually entered commands but it is even hard
to get the commands themselves using the command line history by
pressing Up and Down because auto-issued commands are saved there too.