Refactor GitCommands.{h,cpp}
Factor out the handling of Git subprocess to a single function and remove
the GitCommands-instance-global QProcess instance.

The logic behind invoking the Git subprocess is quite regular. On the
other hand, the process object instance seems sometimes being not fully
de-initialized, so, after several invocations, the waitForFinished()
function call deadlocks and waits for a timeout of 30s.