Updates for source-code FMUs
Make it possible to compile source-code FMUs without setting include-
directives to the preprocessor. This required changes to most files
since they assume preprocessor flags were set.

The generated FMU now contains a list of the files to compile (in the
XML structure), in case the user does not have access to a make system.

Still missing is a config-file containing number of linear systems, etc.
This would make it possible to compile without configure-script at all.
The configure script should also look for the standard FMI includes
which should not need to be present in a source-code FMU.

Use a settings file for FMUs instead of preprocessor, stored inside the
generated FMU.

This allows automatic compilation of the sources, by a tool agnostic of
OpenModelica.

Belonging to [master]:
  - OpenModelica/OMCompiler#2989
  - OpenModelica/OpenModelica-testsuite#1149