Make the diff algorithm handle within in new code
The code assumed that we would not be moving the code where it is
diffed, so the within sort of had to match. Now we add the within
in the source text and it won't show up in the diff but the white-
space should be properly preserved.

Belonging to [master]:
  - OpenModelica/OMCompiler#2998
  - OpenModelica/OpenModelica-testsuite#1147