Use -Wl,-Bstatic for user libraries in FMUs
FMUs should be statically compiled. This enforces it, but might cause
some FMUs to no longer compile if no static library is available or
dependencies are missing in the Library annotation.

Also added a new file Compiler/Util/Autoconf.mo which is generated by
autoconf and contains constants previously in external C functions.

Belonging to [master]:
  - OpenModelica/OMCompiler#3015