[C-Runtime] Initialize with homotopy solver for models with homotopy operator
Use the homotopy method for initialization as default, if the model contains
the homotopy-operator. In that case use 2 homotopy steps in stead of 4 when
doing homotopy after initialization failed without homotopy.
Added new flag "-noHomotopyOnFirstTry" to disable this behaviour. Useful for
models which contains homotopy operator which is "wrong", e.g. some of the
MSL Fluid examples.

  - Fix for ticket #5139.
  - Updated tests which are using homotopy operators and where not initializing
    with homoptopyOnFirstTry.
  - Model from ticket-5206.mos converges to different solution with homotopy,
    since parameter y0 is not fixed.