Add a sanity check that the build folder can be moved
`make install` assumes that `omc` will work even if the installation
directory is moved. This should check that the rpath works fine; it will
not check if files are linked directly against temporary files in the
source directories though.

Changed implementation of getting the installation directory on
Linux by using dladdr like we do on OSX.