remove old unitCheck (#6771)
* [BE] Remove old unit checking module

This is superseded by an NF module with the same name and functionality
and thus no longer needed. Also remove hash tables used by only this
module.

* Handle deprecated flags in a central spot

Right after command line arguments are read and flags are set,
handleDeprecatedFlags detects if a deprecated flag is set and acts
appropriately, so the rest of the code does not need to worry about it.

* [FE] Disable old usage of flag --unitChecking

This flag did some unit checking in the old frontend. Since there is a
new module for that, and the old frontend will be discontinued soon, we
can use this flag for the newer stuff.

* Fix UnitCheck18

* [testsuite] Update tests on unit checking

Removes all old unit checking tests. Tests for the old frontend still
remain in the testsuite because that is not identical to the nf.