Fix mod calculation for integer negative divisor
According to the specification 3.7.1.1 the modulus has the same sign
as the divisor. Unfortunately the % operator in C and C++ works in a
different way. Basically division trunkates towards zero instead of
rounding down towards negative infinity. So we need to manually check
that the result is in the correct range.