[BE] Dont't evaluate inverse functions with noEvent
When a function is inverted inside an if equation with a noEvent
condition, we computed the inverse function anyways, which lead to
problems when arguments were outside the domain of the function.

This is prevented now by having the local auxiliary variables inside
an if block with the corresponding condition.
This should also work for nested conditions.