Change default Cpp solvers (#7684)
* Change Cpp default nonlinear solver to Newton

This is the same solver as used for FMU export.
Some library coverage examples that fail with kinsol work with newton.
The previous default kinsol can still be selected, e.g. through
OMEdit simulation settings or the simflag "-nls=kinsol".

* Change Cpp default linear solver to dgesv with total pivoting

This is the same solver as used for FMU export. The previous default
linearSolver can be selected optionally, in particular if KLU
is enabled for sparse matrices.

* Configure -nls=kinsol for two msl32 tests