Tighten tolerance of continuous Cpp events
This allows to remove a workaround for StateGraph from DASSL.cpp.

The effected examples ExecutionPaths and ShowCompositeStep were working
with 500 steps, but not with 2500 steps in daily library tests anymore.

The change is not applied to OMSI CodegenCpp.tpl, because the combination
cvode/kinsol performs better in msl32_cpp coverage for lose tolerances.