Fix Cpp index of model variables for multidimensional arrays
This is important because states and derivatives are accessed through
lumped vectors z and zDot. Without this fix in particular the mapping
of results to variables for plotting or FMUs is wrong.
See e.g. discretized line models of PowerSystems library.

The fix is not needed for algebraic variables that are accessed through
their individual names.