Improve handling of MSL
Change the order of priority of MODELICAPATH, so it looks in the APPDATA directory first, and then in the omlibrary as a fallback option.
Remove initial splash screen with obsolete library explanation.
Do not load any library by default.
Do not load OpenModelica. We have a nice online documentation for the scripting API.
Improved the libraries settings dialog to add system libraries to load startup.
Allow changing the ModelicaPath