Add basic rules to propagate nominal values
Some simple analysis can be done assuming `a = A*as`, where `A` is
the nominal value of `a` so that `abs(as)` is assumed to be close
to 1 most of the time.

The case `a-b` can be numerically ill conditioned when `A=B` and
`sign(as) = sign(bs)`. In that case the nominal value will be
overestimated. Maybe there is a smart way to fix this?

TODO derive rules for builtin functions.