Fix slow marking of nonlinear iteration vars
Fixes #11302

For each SCC of the system all variables were traversed, so the time
complexity was O(N*S) where N is the number of variables and S is
the number of SCCs.
By first collecting all nonlinear iteration vars in the same set and
then marking them once it should now be O(N).