OMEdit fixes 1.22.1 (#11655)
* fix OMSimulator PYTHONPATH in OMEdit (#11526)

* update subModule OMSimulator

* Check if we have a parameter modification or a binding (#11599)

Fixes #11595

* Do not use elided text for multiline string with fixed font size (#11610)

Fixes #8383

* Add loadClassContentString API (#11574)

* Rename conflicting elements in loadClassContentString (#11613)

Fixes #11596

* Added a tab widget at the bottom of MainWindow (#11627)

* Added a tab widget at the bottom of MainWindow

The tab widget mimics the Messages Browser. It is a shorter version of it.
Only shown when Messages Browser is hidden.
Messages Browser can now only be docked at the bottom.

Scroll to bottom when a new simulation output is added.

Fixes #9889

* Set the index to 0

* Fix the context menu for shapes (#11634)

This is a regression introduced in afb86a3e4c6

* Use the `loadClassContentString` API for cut, copy and paste (#11598)

Instead of using the OMEdit internal structures for pasting elements use `loadClassContentString` to merge.
Generate the Modelica code from the selected elements and then merge the code using `loadClassContentString`.
Put the data in the form of string and JSON on the clipboard. Use it when paste is called.

* Show the typed dimensions of array parameter in the dialog (#11629)

* Improve loadClassContentString (#11644)

- Traverse equations/algorithms manually when renaming instead of using
  the AbsynUtil traversal functions, since the traversal functions only
  traverses expressions and not crefs in e.g. connect equations.
- Rewrite the conflict handling to avoid generating new names that
  conflicts with the elements being added.

* Do not call show documentation when editing documentation

Fixes #11638
Calling show documentation while editing causes a recursion

* Do not push intermediate commands to stack

* Use QScrollArea for welcome page buttons

Fixes #10235

---------

Co-authored-by: arun3688 <rain100falls@gmail.com>
Co-authored-by: Per Östlund <per.ostlund@liu.se>