Get constants through SimCode and Templates (#13248)
Not sure why they still pop up. Perhaps some replacements are not done
fully on function arguments.
This is sort of a hack. Maybe we can fix it in the backend already but I
guess this is more relevant for the NB than the OB and we can close some
open issues with this.