[Codegen] Fix algebraic loops in Jacobian (#14648)
Use hash table of jacobian via contextCref and add test case which checks if these variables are found by the jacobian. The NB has all loops as nonlinear so the nonlinear pipeline is used, which was not used in the OB since algebraic loops in the Jacobian are actually always linear but the NB doesn't know that yet.

---------

Co-authored-by: phannebohm <phannebohm@hsbi.de>