Show the full comment in tooltip
#11721
Auto add ellipses to comment and show full text in tooltip.
Show tooltip for combobox items with full text. ComboBox items are show with ellipses
Do not stretch the fields to show the entire value as this adds the minimum width and then it is not possible to resize the dialog.